Improve point-2d by moving the 'center' parameter.
In previous design, the meaning of the 'center' parameter in the
point-2d struct was ambiguous and difficult to understand. I thought
this is because it is related to a size of an entity not to a position.

So I reinterpreted the point-2d as a representer of a (global or local)
coorination. It requires only a vector from an orign and a rotation.
So I moved the 'center' parameter from it to other components (Now, such
component is only 'model-2d').